Moved item creation and folder methods to vault.py
