+++ app/__init__.py
@@ -47,11 +47,12 @@ def create_app(test_config=None):
     db.init_app(app)
 
     from . import auth
+
     app.register_blueprint(auth.bp)
 
     from . import vault
-    app.register_blueprint(vault.bp)
 
+    app.register_blueprint(vault.bp)
 
     # app = Flask(__name__)
     # db_file = "CAPSTONE-PROJECT.db"
@@ -87,51 +88,6 @@ def create_app(test_config=None):
     def hello():
         return "Hello, World!"
 
-    
-
-    @app.route("/new_folder")
-    def new_folder():
-        return render_template("new_folder.html")
-
-    @app.route("/add_folder", methods=["POST"])
-    def add_folder():
-        folder_name = request.form.get("folder_name")
-
-        messages = []  # List to store messages
-        message_type = "error"  # Default message type
-        conn = db.get_db()
-
-        try:
-            # Check if folder name is provided
-            if not folder_name:
-                messages.append("Please provide a folder name.")
-            else:
-                # Retrieve user ID from session
-                user_id = session.get("user_id")
-                if user_id:
-                    # Add folder to the database with the associated user_id
-                    conn.execute(
-                        "INSERT INTO FOLDER (USER_ID, FOLDER_NAME) VALUES (?, ?)",
-                        (user_id, folder_name),
-                    )
-                    conn.commit()
-                    messages.append("Folder added successfully.")
-                    message_type = "success"
-                else:
-                    messages.append("User ID not found.")
-
-        except Exception as e:
-            messages.append("Failed to add folder. Try again!")
-            print("Error:", e)
-
-        finally:
-            # Close the database connection
-            if conn:
-                conn.close()
-
-        # Redirect back to the vault page after adding the folder
-        return redirect(url_for("vault"))
-
     @app.route("/generate_password", methods=["POST"])
     def handle_generate_password():
         length = int(request.form["total_length"])
@@ -150,7 +106,6 @@ def create_app(test_config=None):
             avoid_ambiguous,
         )
         return password  # Return the generated password as plain text
-
 
     @app.route("/", methods=["GET", "POST"])
     def index():
@@ -185,61 +140,11 @@ def create_app(test_config=None):
     def gen_Password():
         return render_template("gen_Password.html")
 
-
     @app.route("/settings")
     def settings():
         return render_template("settings.html")
 
-    @app.route("/new_item")
-    def new_item():
-        return render_template("new_item.html")
-
-    @app.route("/new_itemAction", methods=["GET", "POST"])
-    def new_itemAction():
-        if "user_id" in session:
-            user_id = session.get("user_id")
-            conn = db.get_db()
-            if request.method == "POST":
-                
-                try:
-                    # Handle form submission
-                    item_type_id = request.form["item_type_id"]
-                    name = request.form["name"]
-                    folder_id = request.form["folder_id"]
-
-                    #conn = database.connect(db_file)
-                    #cursor = conn.cursor()
-
-                    conn.execute(
-                        "INSERT INTO items (item_type_id, name, folder_id, user_id) VALUES (?, ?, ?, ?)",
-                        (item_type_id, name, folder_id, user_id),
-                    )
-                    conn.commit()
-
-                    if conn:
-                        conn.close()
-
-                    return redirect(url_for("success_page"))
-
-                except Error as e:
-                    print("Database Error:", e)
-                    # Handle the error appropriately, e.g., render an error page
-
-            # Retrieve folders belonging to the logged-in user
-            #conn = database.connect(db_file)
-            #cursor = conn.cursor()
-            conn.execute(
-                "SELECT FOLDER_NAME FROM FOLDER WHERE USER_ID = ?", (user_id,)
-            )
-            folders = conn.fetchall()
-            print("Folders:", folders)
-
-            conn.close()
-
-            return render_template("new_item.html", folders=folders)
-        else:
-            # Redirect to login page or handle unauthorized access
-            return redirect(url_for("login_page"))
+    
 
     # Add this route to handle account deletion// not deleting data related to user just user profile
     @app.route("/delete_account", methods=["POST"])
@@ -250,9 +155,9 @@ def create_app(test_config=None):
             conn = db.get_db()
             try:
                 # Connect to the database
-                #conn = database.connect(db_file)
-                #cursor = conn.cursor()
-
+                # conn = database.connect(db_file)
+                # cursor = conn.cursor()
+                # TODO User should have to reauthenticate before deletion
                 # Delete user's data from related tables
                 conn.execute("DELETE FROM REGISTRATION WHERE USER_ID = ?", (user_id,))
                 # You may need additional delete operations for related tables, such as items, folders, etc.
@@ -279,48 +184,7 @@ def create_app(test_config=None):
 
     # Update your settings HTML template to include a form or button to trigger the account deletion
 
-    @app.route("/vault/<folder_name>")
-    def view_folder(folder_name):
-        conn = db.get_db()
-        try:
-            # Establish database connection
-            #conn = database.connect(db_file)
-            #cursor = conn.cursor()
-
-            # Retrieve items from the selected folder
-            conn.execute(
-                "SELECT * FROM ITEM WHERE FOLDER_ID = (SELECT ID FROM FOLDER WHERE LOWER(FOLDER_NAME) = LOWER(?))",
-                (folder_name,),
-            )
-            items = conn.fetchall()
-
-        except database.Error as e:
-            print("Database Error:", e)
-            items = []
-
-        finally:
-            # Close the database connection
-            if conn:
-                conn.close()
-
-        # Render the template with the items in the folder
-        return render_template("folder.html", folder_name=folder_name, items=items)
-
-    @app.route("/get_folders1")
-    def get_folders():
-        # Assuming you have a function to retrieve folders from the database
-        folders = (
-            database.get_folders()
-        )  # Implement this function according to your database schema
-
-        # Assuming each folder is represented as a dictionary with 'folder_id' and 'folder_name' keys
-        folders_data = [
-            {"folder_id": folder["folder_id"], "folder_name": folder["folder_name"]}
-            for folder in folders
-        ]
-
-        # Return the folder data as JSON
-        return jsonify(folders_data)
+    
 
     return app
 
@@ -360,5 +224,5 @@ def generate_password(
     return password
 
 
-#if __name__ == "__main__":
+# if __name__ == "__main__":
 #    app.run(debug=True)
