+++ app/templates/vault.html
@@ -96,9 +96,9 @@
 
 
     </div>
-<a href="/new_item">Create New Item</a>
-        <a href="/new_folder">Create New folder</a>
-        <a href="/profile">Back</a>
+<a href="/vault/new_item">Create New Item</a>
+        <a href="/vault/new_folder">Create New folder</a>
+        <a href="/vault/profile">Back</a>
 
 </body>
 </html>
